use LogMessage instead of BuildLoggerLogLines & refactor
