+++ graphql/util.go
@@ -1,17 +1,11 @@
 package graphql
 
 import (
-	"bufio"
 	"context"
 	"fmt"
-	"io"
 	"net/url"
-	"strconv"
-	"strings"
 	"time"
 
-	"github.com/evergreen-ci/evergreen"
-	"github.com/evergreen-ci/evergreen/apimodels"
 	"github.com/evergreen-ci/evergreen/model"
 	"github.com/evergreen-ci/evergreen/model/artifact"
 	"github.com/evergreen-ci/evergreen/model/task"
@@ -19,11 +13,6 @@ import (
 	restModel "github.com/evergreen-ci/evergreen/rest/model"
 	"github.com/evergreen-ci/evergreen/rest/route"
 	"github.com/evergreen-ci/gimlet"
-	"github.com/evergreen-ci/timber/fetcher"
-	"github.com/mongodb/grip"
-	"github.com/mongodb/grip/level"
-	"github.com/mongodb/grip/message"
-	"github.com/pkg/errors"
 )
 
 // GetGroupedFiles returns the files of a Task inside a GroupedFile struct
@@ -132,98 +121,4 @@ func GetBaseTaskStatusesFromPatchID(r *queryResolver, patchID string) (BaseTaskS
 		baseTaskStatusesByDisplayNameByVariant[task.BuildVariant][task.DisplayName] = task.Status
 	}
 	return baseTaskStatusesByDisplayNameByVariant, nil
-}
-
-func readBuildloggerToSlice(ctx context.Context, taskID string, r io.ReadCloser) []apimodels.LogMessage {
-	lines := []apimodels.LogMessage{}
-	lineChan := make(chan apimodels.LogMessage, 1024)
-	go readBuildloggerToChan(ctx, taskID, r, lineChan)
-
-	for {
-		line, more := <-lineChan
-		if !more {
-			break
-		}
-
-		lines = append(lines, line)
-	}
-
-	return lines
-}
-
-func readBuildloggerToChan(ctx context.Context, taskID string, r io.ReadCloser, lines chan<- apimodels.LogMessage) {
-	var (
-		line string
-		err  error
-	)
-
-	defer close(lines)
-	if r == nil {
-		return
-	}
-
-	reader := bufio.NewReader(r)
-	for err == nil {
-		line, err = reader.ReadString('\n')
-		if err != nil && err != io.EOF {
-			grip.Warning(message.WrapError(err, message.Fields{
-				"task_id": taskID,
-				"message": "problem reading buildlogger log lines",
-			}))
-			return
-		}
-
-		severity := int(level.Info)
-		if strings.HasPrefix(line, "[P: ") {
-			severity, err = strconv.Atoi(strings.TrimSpace(line[3:6]))
-			if err != nil {
-				grip.Error(message.WrapError(err, message.Fields{
-					"task_id": taskID,
-					"message": "problem reading buildlogger log line severity",
-				}))
-				err = nil
-			}
-		}
-
-		select {
-		case <-ctx.Done():
-			grip.Error(message.WrapError(ctx.Err(), message.Fields{
-				"task_id": taskID,
-				"message": "context error while reading buildlogger log lines",
-			}))
-		case lines <- apimodels.LogMessage{
-			Message:  strings.TrimSuffix(line, "\n"),
-			Severity: apimodels.GetSeverityMapping(severity),
-		}:
-		}
-	}
-}
-
-func GetBuildloggerLogs(ctx context.Context, buildloggerBaseURL, taskId, logType string, tail, execution int) (io.ReadCloser, error) {
-	usr := gimlet.GetUser(ctx)
-	fmt.Println(usr.GetAPIKey())
-	taskId = "evergreen_lint_lint_service_d7550a92b28636350af9d375b0df7e341477751d_20_03_06_13_55_56"
-	execution = 0
-	buildloggerBaseURL = "cedar.mongodb.com"
-	opts := fetcher.GetOptions{
-		BaseURL:       fmt.Sprintf("https://%s", buildloggerBaseURL),
-		UserKey:       "1dc6f93f6cde1ea0f4a41b90f36a25af",
-		UserName:      "arjun.patel",
-		TaskID:        taskId,
-		Execution:     execution,
-		PrintTime:     true,
-		PrintPriority: true,
-		Tail:          tail,
-	}
-	switch logType {
-	case apimodels.TaskLogPrefix:
-		opts.ProcessName = evergreen.LogTypeTask
-	case apimodels.SystemLogPrefix:
-		opts.ProcessName = evergreen.LogTypeSystem
-	case apimodels.AgentLogPrefix:
-		opts.ProcessName = evergreen.LogTypeAgent
-	}
-
-	logReader, err := fetcher.Logs(ctx, opts)
-	return logReader, errors.Wrapf(err, "failed to get logs for '%s' from buildlogger, using evergreen logger", taskId)
 }
